backend/requirements.txt
django==2.0
djangorestframework

django-cors-headers
